cloud_run/genai-api/pyproject.toml
[project]
name = "genai-api"
version = "0.1.0"
description = "GenAI API for StyleCo Customer Experience Platform"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.124.0",
    "uvicorn>=0.38.0",
    "google-genai>=1.54.0",
    "jinja2>=3.1.6",
    "pydantic-settings>=2.12.0",
    "google-cloud-aiplatform>=1.119.0",
]
